Crate domain_client_consensus_relay_chain

The relay-chain provided consensus algorithm for parachains.

This is the simplest consensus algorithm you can use when developing a parachain. It is a permission-less consensus algorithm that doesn’t require any staking or similar to join as a collator. In this algorithm the consensus is provided by the relay-chain. This works in the following way.

Each node that sees itself as a collator is free to build a parachain candidate.

This parachain candidate is send to the parachain validators that are part of the relay chain.

The parachain validators validate at most X different parachain candidates, where X is the total number of parachain validators.

The parachain candidate that is backed by the most validators is chosen by the relay-chain block producer to be added as backed candidate on chain.

After the parachain candidate got backed and included, all collators start at 1.

Structs

Verifier

A verifier that just checks the inherents.